requirements.txt
asgiref==3.3.4
Django==2.2.10
djangorestframework==3.12.4
drf-jwt==1.19.1
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
requests==2.25.1
flake8==3.9.2
django-cors-headers==3.7.0
drf-yasg2==1.19.4

